move to utils file to fix build error
