+++ platform/flowglad-next/public/preview/manifest.json
@@ -2,5 +2,5 @@
   "hash": "1b2e536b",
   "path": "/preview/preview.css",
   "size": 87337,
-  "generatedAt": "2025-12-03T03:50:49.792Z"
+  "generatedAt": "2025-12-09T15:11:19.549Z"
 }
